requirements.txt
flask==2.0.1
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
pandas==1.3.3
numpy==1.21.2
validators==0.20.0
openpyxl==3.0.9
reportlab==3.6.1
gunicorn==21.2.0
huggingface-hub==0.20.3
supabase==1.0.3
python-jose==3.3.0
passlib==1.7.4 